Move FieldDescription related classes to FieldDescription directory
