+++ src/FieldDescription/FieldDescriptionFactory.php
@@ -18,7 +18,6 @@ use Doctrine\ODM\MongoDB\Mapping\ClassMetadata;
 use Doctrine\Persistence\ManagerRegistry;
 use Sonata\AdminBundle\Admin\FieldDescriptionInterface;
 use Sonata\AdminBundle\FieldDescription\FieldDescriptionFactoryInterface;
-use Sonata\DoctrineMongoDBAdminBundle\Admin\FieldDescription;
 
 final class FieldDescriptionFactory implements FieldDescriptionFactoryInterface
 {
